+++ Editor/AssetModification/UISoftMaskModifierRunner.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Coffee.UISoftMaskInternal.AssetModification;
+using UnityEditor;
 
 namespace Coffee.UISoftMask
 {
@@ -15,6 +16,7 @@ namespace Coffee.UISoftMask
                     componentModifiers = new IComponentModifier[]
                     {
                         new SoftMaskComponentModifier_Softness(),
+                        new SoftMaskComponentModifier_Alpha(),
                         new SoftMaskComponentModifier_PartOfParent(),
                         new SoftMaskableComponentModifier()
                     }
@@ -25,6 +27,7 @@ namespace Coffee.UISoftMask
                     componentModifiers = new IComponentModifier[]
                     {
                         new SoftMaskComponentModifier_Softness(),
+                        new SoftMaskComponentModifier_Alpha(),
                         new SoftMaskComponentModifier_PartOfParent(),
                         new SoftMaskableComponentModifier()
                     }
@@ -32,6 +35,17 @@ namespace Coffee.UISoftMask
                 (".shader", SoftMaskableShaderModifier.Create)
             })
         {
+        }
+
+        public override void Run(string[] assetPaths, bool dryRun)
+        {
+            if (!dryRun)
+            {
+                AssetDatabase.ImportAsset("Packages/com.coffee.softmask-for-ugui/Shaders",
+                    ImportAssetOptions.ImportRecursive);
+            }
+
+            base.Run(assetPaths, dryRun);
         }
     }
 }